+++ src/Handler/EasyHandle.php
@@ -51,6 +51,11 @@ final class EasyHandle
     public $errno = 0;
 
     /**
+     * @var string|null Effective CURLOPT_PROXY value the handle was created with (if any)
+     */
+    public $effectiveProxy;
+
+    /**
      * @var \Throwable|null Exception during on_headers (if any)
      */
     public $onHeadersException;